Refactor price-finder to only necessary logic
With the sites in place, we can now remove the site code from
price-finder and leave instead just the findItemPrice() and
findItemDetails() functions.  The sites are then responsible for
returning the price, category, and name for the item.

Include tests for these functions, and because price-finder uses
request to scrape the page for price data, include rewire which allows
us to mock out these requests and test the possible returns (without
actually making HTTP requests).