Split out unit and e2e tests
Move the current tests to test/unit in order to add an "e2e" section
which hosts the "end to end" tests.  These tests shouldn't necessarily
be run on every build, but instead to test the integration of actual
requests with the price-finder library.

Also add an npm script goal "test-e2e" to run these tests.