Export the PriceFinder class instead of instance
To allow the user to configure PriceFinder, export the class rather
than the instance.  The user can then pass in an options object which
specifies configuration.  Also add a debug parameter which enables
debugging the PriceFinder object.

Update the tests to account for the change.  Test the debug parameter.

Update the readme to include this information, and additional updates
while we're at it.