Improve error handling for BestBuy site
To avoid the case that the user does not have an API key, and they
request a price for a BestBuy URL, the error that previously was
returned concerned not having a URI to make the page request.  This was
because the API key lookup was within the `getURIForPageData` function.

This commit moves that into the constructor of the BestBuy site, so
that if the API key is not available on instantiation, we throw an
error and include within it the reason why (no API key).  The
`getURIForPageData` function can then generate the URI without worrying
if the API key exists or not.

Also, change the logic so that the environment variable overrides the
configuration.  This allows us the ability to change things in
production without a code change.  Mark this as so in the readme.