Fix for Amazon books accordion
I’d really rather not include the accordion ID tag in the selector for
the books, but Amazon has the same HTML structure for the used and new
book price.  So include the new ID tag in the selector, and some
additional logging to help determine the issue in the future.

This should close #21.