Support additional currencies (Amazon)
What prompted this was issue #19, which requested Amazon support for
additional sites other than “.com”.  In testing of the amazon.fr site,
the main check done in `isSite` failed because of the restriction for
“.com”.  Loosening this allowed price finder to locate the correct
name, but unable to locate the price because of the currency difference.

This commit adds the `accounting` module which gives us the ability to
pull the number value from a variety of currencies.  We’ve pulled the
logic out of `amazon.js` and into a helper library, `site-utils.js`.
Within here, perform an initial check to see which currency type we’re
looking at, and then use `accounting` to find the number value, and
return it to the user.

With this change the test url provided in the issue returns the correct
name and number value for price, but the category is not found.  This
seems like a good first step though…