Update to ES6, apply style guide
This commit is a large scale refactor in updating the content to align
with AirBnb’s ESLint style guide.  This style guide, among other
things, enforces more of the ES6 syntax now supported in Node v4.
We’re also switching to use `class`es for each Site.

Unfortunately, because of things like a change from double quotes to
single quotes, and 4 tab spaces to 2 tab spaces, most all files need
changes.  However, there was no real logic change made in this commit
but instead mostly formatting and style changes.

Because of the upgrade to ES6 syntax, we are removing support for Node
v0.10 and v0.12.

Other things included in this change:
* Add `.editorconfig` and a switch from jshint to ESLint.
* Add `.jscsrc` to provide auto-formatting with tools like Atom.
* Specify the project only supports Node v4 via the `engines` flag in
the `package.json`.
* Remove v0.10, v0.12, and ions from Travis CI file.
* Update dependencies to latest
* Certain tests mock library dependencies, which currently force us to
use `let` in places where we would normally use a `const` type variable.
* The end-to-end tests seem to just have long lines, so disallow
max-len for those tests.
* Readme updates to align with changes