Support find(), improve Steam query
The Steam query requires that we limit the scope of the inner queries
(for the price) by an outer scope (the first game panel display).  To
do this, we’ll first query the outer scope, find the first element, and
then pass that to the `siteUtils.findContentOnPage` to perform the
inner queries for price.

With this new requirement, we may be passing the original jQuery object
or the inner first() one described above to the `findContentOnPage`
function.  So update the function to check if the `find()` function is
available on the jQuery object, and if so, use that.  Otherwise we
proceed as we normally would.