Fix Newegg end to end test: findPriceOnPage
This problem was first found because the end to end tests for Newegg
failed, in that they couldn’t find the price on the site.  After some
investigation, it appeared that the price was not on the initial HTML
page returned from the request call, but perhaps later added to the
page?  What’s weird is that this might have worked in the past?  Unsure…

Either way the current method wasn’t working.  Looking at the HTML
returned from the request, it looks like the price _is_ available in
some javascript near the bottom of the page.  We can get access to this
price via a regular expression.

Update the tests to align with these changes.

This should close #49.