E2E tests: refactor and add to Travis CI
Two general things were done in this commit around the end to end tests:
- Refactor the tests to pull out common function into a
`test-helper.js` file.  This includes the environment variable setup
for Jasmine, creating a single price-finder instance, and defining a
set of helper tests.
- Add the end to end tests to the Travis CI build.  In order to avoid
problems, increase the timeout to a _really_ high value, and set the
retry sleep time to a high value (to avoid spamming).  We also had to
include the key to access Best Buy in an encrypted format within the
.travis.yml file.

According to this document:
https://docs.travis-ci.com/user/speeding-up-the-build/#Parallelizing-you
r-builds-across-virtual-machines
We should be able to have both the unit and e2e tests run in parallel,
and show the unit pass/fail before we see the output of the e2e tests.