Switch from request to superagent
Replace the node module `request` with `superagent`.  Doing so seems to
have greatly improved the reliability of the responses as well as the
detail returned (at least in tests of the existing site pages).  We’re
no longer seeing all the 503 responses (which were seen a lot on
Amazon) and we’re able to get additional page content which makes some
of the sites price queries easier.

In issue #63 a few sites were listed as tests to see if we could
improve the page response data enough to help those example pages.
Newegg still seems to mask their price data too much to allow a single
request to return the price as it is displayed on a rendered page.
Upon further investigation, if one was to view the page source of a
Newegg site, the price data is not included.  This leads one to believe
the price data is loaded on the page after the initial load with
JavaScript.

The test for flipkart had more success, returning the price data
through a simple query after the page load
(`$("*[itemprop='price']").attr('content’)`).

A downside to this implementation is that request headers are not as
simple as they were with the `request` module.  However, since none of
the sites were utilizing this feature, and the flipkart site can use
scraping directly, I would rather include this when required rather
than build functionality without a customer.  This is a breaking change
though.  But I feel the pros (read above) greatly outweigh this con.