Best Buy supports both scraping and API
Refactor the Best Buy site implementation to support both the API path
and the scraping path.  If the API key is available as an environment
variable, we’ll use the API to query for product details.  If however
it’s not available, then we support scraping the site for the details
as well.  This should allow for both those who want to use the API but
not restrict those who do not have an API key.

A breaking change is that we no longer support supplying the API key
through the configuration of the `PriceFinder` object.  To simplify
things, this API key can now only be supplied through an environment
variable (which most should use anyway to avoid committing the API key
to source control).

Update the tests to account for the changes.  The unit tests check both
the API and scraping path.  The end to end tests will check the API
path if the API key is available.  It will always check the scraping
path.  This was done to avoid problems when the API key is not
available for the test user (especially for TravisCI pull requests on
forks who do not have access to the API key).