Standardize processPrice
Since we `toLowerCase` the price string at the start of the function,
we shouldn’t need to `toUpperCase` the string in certain circumstances
to make sure it’s in the correct case — we know it’s in lower case.  So
remove the upper case logic.

Also add test cases for areas not tested.

This should close #81.