Defaults for no currency symbol, fix amazon video games
Finding the price for a video game at Amazon failed because they
changed the page content to remove the decimal point from the price.
To work around this, we’re now only finding the dollar price and not
the cents price.  More information can be found in #97.

This has a side effect of no longer including the currency symbol in
the price which we ask `site-utils` to process.  So update `site-utils`
to check to see if the price string it gets is a number, and if so, use
the defaults in `accounting` to process.  Otherwise, if it’s not a
number, continue to return `-1`.