Flipkart problems
At this point we’re unable to get the price from a Flipkart site.  For
more information on the problem, see:
https://github.com/dylants/price-finder/issues/98
https://github.com/dylants/price-finder/issues/106

So for now, we’re going to state we do NOT support Flipkart, though
we’re leaving in the code until we can determine if we should remove it
or try to add support back in.