Update dependencies / eslint changes
Update the dependencies to the latest versions.  In doing so, there’s a
new version of Airbnb’s eslint rules.  Along with that change comes a
whole set of things we need to do.  Some of those things just don’t
match with the coding style used in this project so far, so some new
rules were added to the `.eslintrc` file.  Otherwise, make the changes
to the source files.