Amazon price fixes, e2e fixes
We’ve collected a list of selectors to use for Amazon products. Go
through that list, trim it down if possible, and label where each
selector is used (at least one place).

In addition to cleaning up the selectors, try to fix a couple issues:
- Use the “more buying options” price when Amazon doesn’t list the
price or doesn’t sell it themselves. This should close #112.
- The second issue is when Amazon has a sale price but doesn’t list it
in the buy box. Move up the sale price selector higher in the select
order to correctly grab it rather than some other price listed in the
buy box (insurance, warranty, etc). This should close #113.

Also fix some end to end tests which require changes because of product
changes on these sites.